feat: create TagsInput component & minor changed (#3376)
* fix: fix credential forms css


* feat: create TagsInput component


* feat: apply tagsInput component to User, serviceAccount page


* fix: fix minor css & text


---------
