+++ apps/web/src/services/asset-inventory/helpers/dynamic-ui-schema-generator/dynamic-layout-schema-template.ts
@@ -58,7 +58,7 @@ export const getDefaultDetailSchema = (fields: DynamicField[], isTrustedAccount:
             }]) || []),
             {
                 key: 'created_at',
-                name: 'Added',
+                name: 'Created',
                 type: 'text',
             },
             {
