+++ apps/web/src/services/iam/components/UserManagementAddTag.vue
@@ -2,29 +2,22 @@
 import {
     reactive, watch,
 } from 'vue';
-import type { TranslateResult } from 'vue-i18n';
 
-import { PFieldGroup, PTextButton, PTextInput } from '@spaceone/design-system';
-import type { InputItem } from '@spaceone/design-system/types/inputs/input/text-input/type';
+import { PTextButton } from '@spaceone/design-system';
 
 import type { Tags } from '@/schema/_common/model';
-import { i18n } from '@/translations';
 
-import { useProxyValue } from '@/common/composables/proxy-state';
-
-import { getInputItemsFromTagKeys } from '@/services/iam/composables/tag-data';
+import TagsInput from '@/common/components/inputs/TagsInput.vue';
 
 interface Props {
     tags: Tags;
     isBordered?: boolean;
-    isEdit?: boolean;
     isFormVisible?: boolean;
 }
 
 const props = withDefaults(defineProps<Props>(), {
     tags: undefined,
     isBordered: false,
-    isEdit: false,
     isFormVisible: false,
 });
 
@@ -33,51 +26,15 @@ const emit = defineEmits<{(e: 'update:tags', role: Tags): void,
 
 const state = reactive({
     formVisible: false,
-    searchText: '',
-    proxyTags: useProxyValue('tags', props, emit),
-    selected: getInputItemsFromTagKeys(props.tags),
 });
-const validationState = reactive({
-    isValid: true as undefined | boolean,
-    invalidText: '' as TranslateResult | string,
-});
+
+const handleUpdateTags = (tags: Tags) => {
+    emit('update:tags', tags);
+};
 
 /* Component */
 const handleClickButton = () => {
     state.formVisible = true;
-};
-const handleChangeInput = (event) => {
-    state.searchText = event.target.value;
-    validationState.isValid = true;
-    validationState.invalidText = '';
-};
-const handleEnterKey = () => {
-    if (state.searchText === '') return;
-    if (!state.searchText.includes(':') || state.searchText.split(':').length > 2) {
-        validationState.isValid = false;
-        validationState.invalidText = i18n.t('IAM.ALT_E_TAG_FORMAT');
-        return;
-    }
-    const isExistItem = state.selected.findIndex((item) => item.name === state.searchText);
-    if (isExistItem !== -1) {
-        validationState.isValid = false;
-        validationState.invalidText = i18n.t('IAM.ALT_E_TAG_DUPLICATION');
-        return;
-    }
-    state.selected.push({ label: state.searchText, name: state.searchText });
-    handleUpdateSelected();
-    validationState.isValid = true;
-    validationState.invalidText = '';
-    state.searchText = '';
-};
-const handleUpdateSelected = (items?: InputItem[]) => {
-    if (items) state.selected = items;
-    const refinedTags = state.selected.map((item) => {
-        const tags = item.name.split(':').map((tag) => tag.trim());
-        return { [tags[0]]: tags[1] };
-    });
-
-    state.proxyTags = Object.assign({}, ...refinedTags);
 };
 
 /* Watcher */
@@ -101,26 +58,9 @@ watch(() => props.isFormVisible, (value) => {
         <div v-else
              class="user-add-tag-form"
         >
-            <p-field-group :label="$t('IAM.USER.FORM.TAGS')"
-                           :invalid-text="validationState.invalidText"
-                           :invalid="!validationState.isValid"
-                           class="input-form"
-            >
-                <p-text-input class="text-input"
-                              multi-input
-                              :invalid="!validationState.isValid"
-                              :selected="state.selected"
-                              appearance-type="stack"
-                              block
-                              @update:selected="handleUpdateSelected"
-                >
-                    <input :placeholder="$t('IDENTITY.TAGS_PLACEHOLDER')"
-                           :value="state.searchText"
-                           @input="handleChangeInput"
-                           @keyup.enter="handleEnterKey"
-                    >
-                </p-text-input>
-            </p-field-group>
+            <tags-input :tags="props.tags"
+                        @update:tags="handleUpdateTags"
+            />
         </div>
     </div>
 </template>
